update project with latest changes

+++ controller/deletecar.php
@@ -2,13 +2,21 @@
 include_once "../config.php";
 
 try {
-    $pdo = new PDO("mysql:host=$host;dbname=$dbname", $username, $password);
+    $pdo = config::getConnexion();
     $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-    $ID_VOITURE = $_GET['id_voiture'];
-    $stmt = $pdo->prepare("DELETE FROM car WHERE id_voiture = :id_voiture");
-    $stmt->bindParam(':id_voiture', $ID_VOITURE);
-    $stmt->execute();
-    header('location: showcar.php');
+    
+    if(isset($_GET['id_voiture'])) {
+        $id_voiture = $_GET['id_voiture'];
+
+        $stmt = $pdo->prepare("DELETE FROM car WHERE id_voiture = :id_voiture");
+        $stmt->bindParam(':id_voiture', $id_voiture);
+        $stmt->execute();
+
+        header('location: showcar.php');
+        exit();
+    } else {
+        echo "Car ID not provided.";
+    }
 } catch(PDOException $e) {
     echo "Deletion failed: " . $e->getMessage();
 }
